Removed SCSS warnings
SCSS was giving warnings about the use of @include.
The only thing being included was Bulma. In the interest of ensuring Bulma was up to date with SCSS' expectations, it was also updated to the latest.
bulma-prefers-dark was also removed, as it is supported by default in Bulma now.
In the interest of getting things updated, the file was set to use the UTF-8 charset, this included adding the UTF-8 BOM to the file.
This does result in a change in UI colors, since some were adjusted. Notably the info color remains bound to cyan, but the definition for cyan was updated from a cyan-blue to be closer to cyan.
There are other minor changes, but overall, the UI is still functional.