Normalize code formatting
The code had non-uniform formatting all throughout. Some blocks had too few indents, some had not enough. As best as I could tell the target style was 2 indents for every block.
There were also some places where types did not have a space after the colon, and others where there was a space, it seemed most likely that the intent was to have a space.
Some type unions had spaces around the pipes, some did not, overall it seemed that the intent was to have spaces.
A couple pieces of HTML had been wrapped oddly, those were corrected to follow a consistent 2 indents. A couple of long lines were also wrapped.
In a couple of places where functions had many parameters that were wrapped, a newline was added after the last parameter, and the closing parenthesis was left-dented to be inline with the column of the beginning of the function declaration.
All newlines before the closing script tag were removed. Newlines were added to the end of every file, some editors behave oddly when there isn't that newline there.