Add custom title generation prompts
Moved the prompt for title generation from the system prompt to a new field.
This is necessary since Jen currently uses itself as its title generator, and thus we need to distinguish between the overall system prompt and the title prompt. This also allows other models to use custom prompts different from the default one for its title generator.