Update title generation to use the profiles specified title generator
This moves from using the current model being used for title generation, to using the specific title generation profile.
The reason for this is that you may be conversing with the very expensive GPT 4.5, or some other preview model, and don't want to be billed so harshly for a simple title.
This also allows for overrides to the title generation prompt, in case the hardcoded one doesn't work well.
The current Title Generator profile uses the cheapest GPT 4.1 nano model, but it's not great at following instructions. I would recommend the GPT 4o mini model, which has similar pricing and better titles. At the current time, this is not existent in the available models though, so we're sticking with GPT 4.1 nano for now.